updated documentation, providing roadmap for project
Modified README to include a step-by-step of the character generation
process. This is more for me than anything, and will likely change in
the future (when you’ll be able to read this).

Created a .old folder for old concepts and ideas, and added it to
gitignore to help keep things tidy